+++ src/components/utdragFraSykefravaeret/Sykmeldinger.tsx
@@ -129,8 +129,10 @@ const SykmeldingTittelbeskrivelse = ({ sykmelding }: UtvidbarTittelProps) => {
           {erViktigInformasjon && (
             <img
               height={18}
+              width={18}
               alt="Viktig informasjon"
               src={MerInformasjonImage}
+              className="max-w-[18px]"
             />
           )}
         </div>
