+++ src/utils/aktivitetskravUtils.ts
@@ -1,9 +1,5 @@
-import {
-  AktivitetskravDTO,
-  AktivitetskravVurderingDTO,
-} from "@/data/aktivitetskrav/aktivitetskravTypes";
+import { AktivitetskravDTO } from "@/data/aktivitetskrav/aktivitetskravTypes";
 import { OppfolgingstilfelleDTO } from "@/data/oppfolgingstilfelle/person/types/OppfolgingstilfellePersonDTO";
-import dayjs from "dayjs";
 
 export const oppfolgingstilfelleForAktivitetskrav = (
   aktivitetskrav: AktivitetskravDTO,
@@ -22,16 +18,4 @@ export const gjelderOppfolgingstilfelle = (
     aktivitetskrav.stoppunktAt > oppfolgingstilfelle.start &&
     aktivitetskrav.stoppunktAt <= oppfolgingstilfelle.end
   );
-};
-
-export const isExpiredForhandsvarsel = (
-  vurdering: AktivitetskravVurderingDTO
-): boolean => {
-  if (vurdering.varsel?.svarfrist) {
-    const svarfrist = dayjs(vurdering.varsel.svarfrist);
-    const today = dayjs(new Date());
-    return svarfrist.isBefore(today, "date");
-  }
-
-  return false;
 };
